# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Card catalog. Admin writes require Authorization: Bearer \u003cADMIN_TOKEN\u003e.",
        "title": "Engmark API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
        "/api/v1/admin/cards": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create card",
                "parameters": [
                    {
                        "description": "card without id and version",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/catalog_transport_http.CreateCardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/catalog_transport_http.CardDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "catalog_transport_http.CardDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deckId": {
                    "type": "integer"
                },
                "example": {
                    "type": "string"
                },
                "exampleHighlight": {
                    "type": "string"
                },
                "exampleRu": {
                    "type": "string"
                },
                "extra": {
                    "type": "string"
                },
                "extraLabel": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ipa": {
                    "type": "string"
                },
                "pos": {
                    "type": "string"
                },
                "posRu": {
                    "type": "string"
                },
                "rusTrans": {
                    "type": "string"
                },
                "stress": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                },
                "translation": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                },
                "word": {
                    "type": "string"
                }
            }
        },
...
        "catalog_transport_http.CreateCardRequest": {
            "type": "object",
            "properties": {
                "deckId": {
                    "type": "integer"
                },
                "example": {
                    "type": "string"
                },
                "exampleHighlight": {
                    "type": "string"
                },
                "exampleRu": {
                    "type": "string"
                },
                "extra": {
                    "type": "string"
                },
                "extraLabel": {
                    "type": "string"
                },
                "ipa": {
                    "type": "string"
                },
                "pos": {
                    "type": "string"
                },
                "posRu": {
                    "type": "string"
                },
                "rusTrans": {
                    "type": "string"
                },
                "stress": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                },
                "translation": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
...
        "core_http_response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "core_http_response.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/core_http_response.ErrorBody"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Value is \"Bearer \u003cADMIN_TOKEN\u003e\". Do not put a real token here.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}